% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_times.R
\name{map_times}
\alias{map_times}
\title{Discretize observation times to a grid of discrete times}
\usage{
map_times(t0, tf, dt, t)
}
\arguments{
\item{t0}{\code{POSIXct} or \code{numeric} for start of time grid}

\item{tf}{\code{POSIXct} or \code{numeric} for end of time grid}

\item{dt}{Number of seconds between time steps}

\item{t}{Vector of times to map to grid}
}
\description{
Discretization maps the input times \code{t} to the closest matching time 
in the regular grid of discrete times implied by \code{t0}, \code{tf}, and
\code{dt}.
}
\examples{
map_times(t0 = 0, tf = 1, dt = .001, t = c(.123, .234, .345))
map_times(
    t0 = round(Sys.time()), 
    tf = round(Sys.time()) + 3600, 
    dt = 1, 
    t = Sys.time() + c(20, 60, 90)
 )

}
